Various Bugfixes (#12296)
* Various bugfixes.
Fixes #8286
Fixes #8416
Fixes #8492
Fixes #9147
Fixes #10166
Fixes #12126
Fixes #12275

* Removes some debug output.

* Changes the beach ball solution.

* Makes the put_in_hand_check() call independent of the get_inactive_hand() call.

* Made a reagent_containers-level proc for displaying a list of reagents in a readable format.

+++ code/modules/mob/living/carbon/alien/special/facehugger.dm
@@ -142,7 +142,7 @@ var/const/MAX_ACTIVE_TIME = 400
 		if(real) // Lamarr still tries to couple with heads, but toys won't
 			processing_objects.Add(src)
 
-	else
+	else if(!sterile)
 		qdel(src)
 
 /obj/item/clothing/mask/facehugger/examine(mob/user)
@@ -244,7 +244,7 @@ var/const/MAX_ACTIVE_TIME = 400
 			if(istype(mouth_protection, /obj/item/clothing/head/helmet/space/rig))
 				rng = 15
 			if(prob(rng)) // Temporary balance change, all mouth-covering hats will be more effective
-				H.visible_message("<span class='danger'>\The [src] smashes against [H]'s [mouth_protection], and rips it off in the process!</span>")
+				H.visible_message("<span class='danger'>\The [src] smashes against [H]'s \the [mouth_protection], and rips it off in the process!</span>")
 				H.drop_from_inventory(mouth_protection)
 				GoIdle(15)
 				return
